Separate line highlighter in dedicated class
