+++ src/report_modules/report_module_gui/CMakeLists.txt
@@ -14,6 +14,7 @@ add_executable(${REPORT_MODULE_GUI_PROJECT}
     src/report_format_ui.cpp
     src/ui/c_checker_widget.cpp
     src/ui/c_report_module_window.cpp
+    src/ui/c_line_highlighter.cpp
     src/qrc/poc__report_module_gui__qrc.cpp
 )
